+++ core/src/main/kotlin/at/bitfire/davdroid/ui/setup/LoginScreenModel.kt
@@ -16,8 +16,6 @@ import at.bitfire.davdroid.repository.AccountRepository
 import at.bitfire.davdroid.servicedetection.DavResourceFinder
 import at.bitfire.davdroid.settings.AccountSettings
 import at.bitfire.davdroid.settings.SettingsManager
-import at.bitfire.davdroid.validation.LoginValidator
-import at.bitfire.davdroid.validation.SyncValidator
 import at.bitfire.vcard4android.GroupMethod
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
