+++ core/src/main/kotlin/at/bitfire/davdroid/ui/setup/LoginValidator.kt
@@ -2,7 +2,7 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
 
-package at.bitfire.davdroid.validation
+package at.bitfire.davdroid.ui.setup
 
 import dagger.BindsOptionalOf
 import dagger.Module
